Main and Camel samples
