+++ .travis.yml
@@ -1,0 +1,20 @@
+language: scala
+
+scala:
+  - "2.11.8"
+  - "2.12.1"
+
+jdk:
+  - oraclejdk8
+
+before_cache:
+  - find $HOME/.ivy2 -name "ivydata-*.properties" -print -delete
+  - find $HOME/.sbt  -name "*.lock"               -print -delete
+
+cache:
+  directories:
+    - $HOME/.ivy2/cache
+    - $HOME/.sbt/boot
+
+script:
+  - sbt -J-XX:ReservedCodeCacheSize=128m ++$TRAVIS_SCALA_VERSION test && mvn test -P $TRAVIS_SCALA_VERSION
