--- akka-sample-camel-java/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-camel-java
-title=Akka Camel Samples with Java
-description=Akka Camel Samples with Java
-tags=akka,camel,java,sample
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
