+++ akka-sample-camel-java/project/plugins.sbt
