--- akka-sample-camel-scala/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-camel-scala
-title=Akka Camel Samples with Scala
-description=Akka Camel Samples with Scala
-tags=akka,camel,scala,sample
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
