+++ akka-sample-camel-scala/build.sbt
@@ -1,12 +1,10 @@
+organization := "com.typesafe.akka.samples"
 name := "akka-sample-camel-scala"
 
-version := "2.4-SNAPSHOT"
-
-scalaVersion := "2.11.7"
-
+scalaVersion := "2.12.1"
 libraryDependencies ++= Seq(
-  "com.typesafe.akka" %% "akka-actor" % "2.4-SNAPSHOT",
-  "com.typesafe.akka" %% "akka-camel" % "2.4-SNAPSHOT",
+  "com.typesafe.akka" %% "akka-actor" % "2.4.14",
+  "com.typesafe.akka" %% "akka-camel" % "2.4.14",
   "org.apache.camel" % "camel-jetty" % "2.13.4",
   "org.apache.camel" % "camel-quartz" % "2.13.4",
   "org.slf4j" % "slf4j-api" % "1.7.2",
