+++ akka-sample-camel-scala/project/plugins.sbt
