--- akka-sample-main-java-lambda/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-main-java-lambda
-title=Akka Main in Java with Lambdas
-description=Actor based version of obligatory Hello World program using the generic launcher class akka.Main.
-tags=Basics,akka,java,java8,starter
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
