--- akka-sample-main-java-lambda/build.sbt
@@ -1,11 +1,0 @@
-name := "akka-sample-main-java-lambda"
-
-version := "2.4-SNAPSHOT"
-
-scalaVersion := "2.11.7"
-
-libraryDependencies ++= Seq(
-  "com.typesafe.akka" %% "akka-actor" % "2.4-SNAPSHOT"
-)
-
-licenses := Seq(("CC0", url("http://creativecommons.org/publicdomain/zero/1.0")))
