--- akka-sample-main-java-lambda/src/main/java/sample/hello/Main.java
@@ -1,8 +1,0 @@
-package sample.hello;
-
-public class Main {
-
-  public static void main(String[] args) {
-    akka.Main.main(new String[] { HelloWorld.class.getName() });
-  }
-}
