--- akka-sample-main-java-lambda/src/main/resources/application.conf
@@ -1,3 +1,0 @@
-akka {
-  loglevel = INFO
-}
