--- akka-sample-main-java/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-main-java
-title=Akka Main in Java
-description=Actor based version of obligatory Hello World program using the generic launcher class akka.Main. 
-tags=Basics,akka,java,starter
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
