+++ akka-sample-main-java/pom.xml
@@ -4,21 +4,63 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>akka-sample-main-java</artifactId>
+  <artifactId>akka-sample-main-lambda</artifactId>
   <groupId>com.typesafe.akka.samples</groupId>
   <name>Akka Main in Java</name>
-  <version>1.0</version>
-  
+  <version>empty</version>
+
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
-  
+
+  <profiles>
+    <profile>
+      <id>2.12.1</id>
+      <activation>
+        <activeByDefault>true</activeByDefault>
+      </activation>
+      <properties>
+        <scalaBinaryVersion>2.12</scalaBinaryVersion>
+      </properties>
+    </profile>
+    <profile>
+      <id>2.11.8</id>
+      <properties>
+        <scalaBinaryVersion>2.11</scalaBinaryVersion>
+      </properties>
+    </profile>
+  </profiles>
+
   <dependencies>
     <dependency>
       <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-actor_2.11</artifactId>
-      <version>2.4-SNAPSHOT</version>
+      <artifactId>akka-actor_${scalaBinaryVersion}</artifactId>
+      <version>2.4.14</version>
     </dependency>
   </dependencies>
 
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.1</version>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+          <fork>true</fork>
+          <compilerArgs>
+            <arg>-Xlint</arg>
+          </compilerArgs>
+        </configuration>
+      </plugin>
+    </plugins>
+    <extensions>
+      <extension>
+        <groupId>fr.brouillard.oss</groupId>
+        <artifactId>jgitver-maven-plugin</artifactId>
+        <version>0.1.1</version>
+      </extension>
+    </extensions>
+  </build>
 </project>
