+++ akka-sample-main-java/project/plugins.sbt
