+++ akka-sample-main-java/src/main/java/sample/hello/HelloWorld.java
@@ -1,25 +1,26 @@
 package sample.hello;
 
+import akka.actor.AbstractActor;
 import akka.actor.Props;
-import akka.actor.UntypedActor;
 import akka.actor.ActorRef;
+import akka.japi.pf.ReceiveBuilder;
+import static sample.hello.Greeter.Msg;
 
-public class HelloWorld extends UntypedActor {
+public class HelloWorld extends AbstractActor {
+
+  public HelloWorld() {
+    receive(ReceiveBuilder.
+      matchEquals(Msg.DONE, m -> {
+        // when the greeter is done, stop this actor and with it the application
+        context().stop(self());
+      }).build());
+  }
 
   @Override
   public void preStart() {
     // create the greeter actor
     final ActorRef greeter = getContext().actorOf(Props.create(Greeter.class), "greeter");
     // tell it to perform the greeting
-    greeter.tell(Greeter.Msg.GREET, getSelf());
-  }
-
-  @Override
-  public void onReceive(Object msg) {
-    if (msg == Greeter.Msg.DONE) {
-      // when the greeter is done, stop this actor and with it the application
-      getContext().stop(getSelf());
-    } else
-      unhandled(msg);
+    greeter.tell(Greeter.Msg.GREET, self());
   }
 }
