--- akka-sample-main-scala/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-main-scala
-title=Akka Main in Scala
-description=Actor based version of obligatory Hello World program using the generic launcher class akka.Main. 
-tags=Basics,akka,scala,starter
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
