+++ akka-sample-main-scala/build.sbt
@@ -1,29 +1,9 @@
-import NativePackagerHelper._
-
+organization := "com.typesafe.akka.samples"
 name := "akka-sample-main-scala"
 
-version := "2.4-SNAPSHOT"
-
-scalaVersion := "2.11.7"
-
+scalaVersion := "2.12.1"
 libraryDependencies ++= Seq(
-  "com.typesafe.akka" %% "akka-actor" % "2.4-SNAPSHOT"
+  "com.typesafe.akka" %% "akka-actor" % "2.4.14"
 )
-
-enablePlugins(JavaServerAppPackaging)
-
-mainClass in Compile := Some("sample.hello.Main")
-
-mappings in Universal ++= {
-  // optional example illustrating how to copy additional directory
-  directory("scripts") ++
-  // copy configuration files to config directory
-  contentOf("src/main/resources").toMap.mapValues("config/" + _)
-}
-
-// add 'config' directory first in the classpath of the start script,
-// an alternative is to set the config file locations via CLI parameters
-// when starting the application
-scriptClasspath := Seq("../config/") ++ scriptClasspath.value
 
 licenses := Seq(("CC0", url("http://creativecommons.org/publicdomain/zero/1.0")))
