+++ pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>akka-samples</artifactId>
+    <groupId>com.typesafe.akka.samples</groupId>
+    <name>Akka Samples</name>
+    <version>empty</version>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>akka-sample-main-java</module>
+    </modules>
+</project>
